requirements.txt
langchain
langchain-openai
langgraph
redis
qdrant-client
python-dotenv